Moved styles dir to out of utils
