+++ src/components/BuildStatus.js
@@ -3,7 +3,7 @@ import { CX_OFF_WHITE, CX_FONT, BATMAN_GRAY } from "../utils/Constants";
 import BuildIcon from "./BuildIcon";
 import { Card, CardContent } from "@material-ui/core";
 import { overviewURL, jenkinsURLList } from "../utils/Link";
-import { BOX_STYLE, BOX_HEADER } from "../utils/styles/styles";
+import { BOX_STYLE, BOX_HEADER } from "../styles/styles";
 import { hour, getRelativeTime } from "../utils/TimeUtils";
 
 const styles = {
