+++ src/components/Github.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Card } from "@material-ui/core";
 import { CX_GRAY_BLUE } from "../utils/Constants.js";
-import { BOX_STYLE, BOX_HEADER } from "../utils/styles/styles";
+import { BOX_STYLE, BOX_HEADER } from "../styles/styles";
 import pullRequest from "../../resources/pullRequest.png";
 import { getRelativeTime, hour } from "../utils/TimeUtils";
