+++ src/components/SlackComponent.js
@@ -2,7 +2,8 @@ import React from "react";
 import SlackCard from "./SlackCard";
 import { CX_OFF_WHITE, CX_FONT } from "../utils/Constants";
 import Card from "@material-ui/core/Card";
-import { BOX_STYLE, BOX_HEADER } from "../utils/styles/styles";
+import Grow from "@material-ui/core/Grow";
+import { BOX_STYLE, BOX_HEADER } from "../styles/styles";
 import { minute, time } from "../utils/TimeUtils";
 import { GITHUB_HEIGHT } from "./Github";
