+++ src/wallboards/TVWallboard.js
@@ -2,7 +2,7 @@ import React from "react";
 import BuildStatus from "../components/BuildStatus";
 import SlackComponent from "../components/SlackComponent";
 import Github from "../components/Github";
-import { leftBox, rightBox } from "../utils/styles/WallboardStyles";
+import { leftBox, rightBox } from "../styles/WallboardStyles";
 import Grid from "@material-ui/core/Grid";
 
 export default class TVWallboard extends React.Component {
